Added global config for menu and header
Added a haeder.js file for the header in the subpages and and a
menu-bar.js file for the menu in the subpages.
The js files are just document.write() commands that writes the string
in each file.
So if we want to make a change or add a new page we just have to change
or include the scripts in the html file respectively and we are good to go.

Also added the photos for the Committees